+++ widget/src/index.ts
@@ -1,4 +1,6 @@
 import { Html, renderHtml, default as HtmlDisplay } from "./htmlDisplay"
+import { type Props as GraphvizDisplayProps, default as GraphvizDisplay } from "./graphvizDisplay"
 
 export type { Html }
 export { renderHtml, HtmlDisplay }
+export { GraphvizDisplayProps, GraphvizDisplay }